POD update -
API/functionality change.

A design decision was made to pass the Krang::Publisher object along as part of the publish process in the element tree.  For this to make sense, a number of read-only accessors need to be made available.

First set of accessors (is_preview(), is_publish(), category()) added.


git-svn-id: http://svn.krangcms.com/trunk@731 689d98c0-9023-0410-9c0b-f79783857903