Updated POD for publish.
Made the decision to pass along the Krang::Publisher object as part of the publish process.

As such, a number of accessors are made available by the Krang::Publisher object that will offer information as to current state.  Krang::Publisher will also return the current category & story objects, so the parameter list for the various subroutines here is greatly simplified.


git-svn-id: http://svn.krangcms.com/trunk/krang@733 689d98c0-9023-0410-9c0b-f79783857903