Krang::Publisher->template_search_path now takes an optional category argument.
If the argument is supplied, it bases the returning path list on the supplied Krang::Category->url().  Otherwise, it uses the currently stored Krang::Category object, corresponding to the current category in a publish run.

The motivation - Krang::Publisher->deploy_template() uses template_search_path to determine where to write a template out to.  In a case like that, there's no current category, as there is no publish run going on.  And if there is a category set, you most likely shouldn't be using it anyways.


git-svn-id: http://svn.krangcms.com/trunk@754 689d98c0-9023-0410-9c0b-f79783857903