According to Terrie:
I have made a few changes to my screens as the ones committed were not
the most recent versions and/or I noticed some problems with validation.
All the screens below can be committed when you get the chance with the
addition of Template/list_active.tmpl.

For some reason I am having trouble viewing Group/edit/view.tmpl, so
there may be a problem with that one.

-Merged with Terrie's new versions.
-Fixed problem with Group/edit_view.tmpl (was not using 'may_checkin_all')


git-svn-id: http://svn.krangcms.com/trunk@1126 689d98c0-9023-0410-9c0b-f79783857903