Templates can now be flagged for testing and used in preview.
preview_story() now checks for templates that are checked out by the current user & flagged as 'testing' - if any exist, they are deployed to a temporary directory.  The tempdir & templates within are removed from the filesystem as part of preview_story() cleanup.

template_search_path() will now return template testing paths as well as deploy paths, but only if the run is a preview, and a template testing path exists for this user.  The end result is that Krang::ElementClass->find_template() did not need to be touched whatsoever to support this feature.

Code that existed in several places to handle the writing of template data to the filesystem has been centralized into a single _write_template() internal method.


git-svn-id: http://svn.krangcms.com/trunk@1135 689d98c0-9023-0410-9c0b-f79783857903