This is probably the largest change in publishing functionality in about 6 months, so keep an eye on things - all tests do pass, for the record.
Krang::ElementClass->fill_template() had a bug where if an element had a child named 'title', that element would not get published in the template, in favor of $story->title().

The actual behavior was that a number of variables (title, page_break, cover_date, content) would use the internally-created values, and ignore any child elements with those names.

The change is that child elements no longer check to see if variables based on their names have already been created, except in the context of other children.  At the end, the template is populated with both sets of variables, but the children will win in all cases where variables were created on both sides.

This change will affect any element that does not override fill_template(), or that still calls SUPER->fill_template().  That being said, it would only arise if that element had children called 'title', 'page_break', 'cover_date', or 'content', and in which case the old results were broken anyways.


git-svn-id: http://svn.krangcms.com/trunk@2266 689d98c0-9023-0410-9c0b-f79783857903