Fixed additional_content_block() to work right!
About time.

Krang::Publisher->additional_content_block() can now be called from any point during the publish process.  It no longer returns formatted output to be integrated in the final output.

Instead, it maintains an internal list of all the additional files to be written out, and as before, writes them out at the end.

Should have done this a long time ago.  Should be backwards-compatible with the old way of doing things, as nothing has really changed in the end.

Updated customizing_publish.pod doc as well to reflect the changes.


git-svn-id: http://svn.krangcms.com/trunk@2683 689d98c0-9023-0410-9c0b-f79783857903