Workaround for a bug in the Imager libraries to keep Krang::Media from segfaulting when attempting to generate thumbnails for certain images.
The problem is that when Imager scales an image, if one dimension ends up less than zero, that dimension gets truncated, and is set equal to zero.  On output, this causes a segfault.

Krang::Media will no longer write out a file with dimensions < 1 on either axis.  This will cause some images (those with extreme aspect ratios, something like a 5x100 image) to not be thumbnailed properly.

This is temporary - the next release of Imager will fix this bug, and all images will then be thumbnailed properly.


git-svn-id: http://svn.krangcms.com/trunk@2689 689d98c0-9023-0410-9c0b-f79783857903