fixed a circular dependency in Krang::Category->deserialize_xml()
Categories that linked to stories or media objects in their element trees, where those stories/media belonged to the same category would fail on import, as the circular reference was not handled properly.

Fixed - on deserialization, the category is saved & registered with the DataSet object before the element tree is then deserialized.


git-svn-id: http://svn.krangcms.com/trunk@2826 689d98c0-9023-0410-9c0b-f79783857903