Patches to the search-preferences functionality.
Reworked how things were stored in the Session hash - search prefs are now stored in a multidimensional structure within the session hash - makes it easier to encapsulate Story/Media/Template prefs.

Fixed the broken "Clear" button in the advanced-search screens.

Fixed improper variable-name pollution in the caching that was leading to ISEs because the 'rm' variable was being set to runmodes that did not exist.


git-svn-id: http://svn.krangcms.com/trunk@2873 689d98c0-9023-0410-9c0b-f79783857903