Fixed tmp/ cleanup in Krang::Schedule.
Made the mistake of assuming that File::Path::rmtree would recursively delete a directory tree in the fashion of rm -rf.  It does not do that - it only deletes directories, and will therefore choke on any dir structure that has files.

The fix - for all directories to be deleted, recurse through using File::Find, deleting any files first.  Problem solved.


git-svn-id: http://svn.krangcms.com/trunk@2874 689d98c0-9023-0410-9c0b-f79783857903