Most of the time we don't want AJAX requests to have anything to do with the navigation since it's almost always going to be the same. But there are a couple of places (DeskAdmin, Group) where it can change based on the user's action.
Add Krang::CGI->update_nav() that will send a signal to the client side
(using a JSON header) that the current request that's being returned modified
the nav and it needs to be updated.

Also add Krang::CGI::Nav and nav.pl which just return the nav for the current
user.


git-svn-id: http://svn.krangcms.com/branches/ajax_and_new_skin@3819 689d98c0-9023-0410-9c0b-f79783857903