The story-preview-link Behavior does not work from the SlideDown alert messages.
This steps on any changes you've made to the story but not saved by taking you 
to the find stories screen in the current window/tab.

Switching this back to a "javascript:Krang.preview()" link, and adding a call
to format_url for wrapping.

If you have suggestions on making the behavior work in the alert message, I'm all ears.


git-svn-id: http://svn.krangcms.com/trunk@6061 689d98c0-9023-0410-9c0b-f79783857903